refactor: move `imgixLoader` to its own file
